Eliminacion En todas las tablas para la interfaz de Administrador : Rosio, Ana, Jesus Guerrero

+++ public_html/Paginas/procesar/eliminar.php
@@ -2,61 +2,54 @@
 <?php 
 require_once 'config.php';
 $conexion =  mysqli_connect(config::$servidor, config::$usuario, config::$password, config::$baseDeDatos);
+if(isset($_GET['ID']) and isset($_GET['me'])){
  $idgeneral=$_GET['ID'];
- $metodoElegido=$_GET['me'];
+ $tabla=$_GET['me'];
 
- if($metodoElegido==1){
- 	EliminarUsuarios();
- }elseif ($metodoElegido==2) {
- 	EliminarProductos();
- }elseif($metodoElegido==3){
-EliminarProveedores();
- }elseif($metodoElegido==4){
-EliminarSugerencias();
- }elseif($metodoElegido==5){
-EliminarVentas();
+ if($tabla==1){
+ 	$Eliminacion=mysqli_query($conexion,"DELETE FROM usuarios WHERE IdCliente='$idgeneral'");
+header('Location:../ConUsuario.php');
  }
-	function EliminarUsuarios(){
-	$Eliminacion=mysqli_query($conexion,"DELETE FROM usuarios WHERE IdCliente='$idgeneral'");
+
+  if($tabla==2){
+ 	$Eliminacion=mysqli_query($conexion,"DELETE FROM productos WHERE IdProducto='$idgeneral'");
+ 	header('Location:../ConProductos.php');
+ }
+
+ if($tabla==3){
+ $Eliminacion=mysqli_query($conexion,"DELETE FROM proveedores WHERE IdProveedor='$idgeneral'");
+ header('Location:../ConProveedores.php');
+ }
+
+ if($tabla==4){
+ 	$Eliminacion=mysqli_query($conexion,"DELETE FROM sugerencias WHERE IdSugerencia='$idgeneral'");
+ 	header('Location:../ConSugerencias.php');
+ }
+
+ if($tabla==5){
+ 	$Eliminacion=mysqli_query($conexion,"DELETE FROM ventas WHERE IdVenta='$idgeneral'");
+ 	header('Location:../ConVentas.php');
+ }
+
+ 
+
+
+
+	
+	function RecargarProductos(){
+	
 	}
-	function EliminarProductos(){
-	$Eliminacion=mysqli_query($conexion,"DELETE FROM productos WHERE IdProducto='$idgeneral'");
+	function RecargarProveedores(){
+	
 	}
-	function EliminarProveedores(){
-	$Eliminacion=mysqli_query($conexion,"DELETE FROM proveedores WHERE IdProveedor='$idgeneral'");
-	}
-	function EliminarSugerencias(){
-	$Eliminacion=mysqli_query($conexion,"DELETE FROM sugerencias WHERE IdSugerencia='$idgeneral'");
+	function RecargarSugerencias(){
+	
 	}
 	function EliminarVentas(){
-	$Eliminacion=mysqli_query($conexion,"DELETE FROM ventas WHERE IdVenta='$idgeneral'");
+	
 	}
-	//$resultado=$conexion->conexion($consulta);
-	
-?>
 
-<html background="black">
-	<head>
-		<title>Eliminar usuario</title>
-	</head>
-	
-	<body>
-		<center>
-			<?php 
-				if($Eliminacion>0){
-				?>
-				
-				<h1>Usuario Eliminado</h1>
-				
-				<?php 	}else{ ?>
-				
-				<h1>Error al Eliminar Usuario</h1>
-				
-			<?php	} ?>
-			<p></p>		
-			
-			<a href="consultar.php">Regresar</a>
-			
-		</center>
-	</body>
-</html>
+	}else{
+		die('La operacion no puede ser realizada, Ya no Existen elementos registrados en: ERROR 0xT');	
+	}
+?>
